chore: setup local development env
