+++ .gitignore
@@ -43,3 +43,5 @@ go.work
 
 # Built Visual Studio Code Extensions
 *.vsix
+
+tmp/
