+++ Dockerfile.dev
@@ -1,0 +1,13 @@
+FROM golang:1.21
+
+WORKDIR /src
+
+RUN go install github.com/cosmtrek/air@latest
+
+COPY go.mod go.sum ./
+
+RUN go mod download
+
+COPY . .
+
+CMD ["air", "-c", ".air.toml"]
