+++ docker-compose.yaml
@@ -1,0 +1,18 @@
+version: '3.1'
+
+services:
+
+  carbon:
+    build:
+      context: .
+      dockerfile: Dockerfile.dev
+    ports:
+      - 8000:8000
+    volumes:
+      - ./:/src
+      # Volume for Google Cloud Credentials
+      - ~/.config/gcloud/application_default_credentials.json:/credentials/application_default_credentials.json
+      # we need to volume in our emissions data factors for now
+      - ../emissions-data:/tmp/emissions-data
+      # TODO volume for AWS credentials
+
